refactor: refactor taskDto and taskMapper, fix style
