+++ src/main/java/ru/mssecondteam/taskservice/dto/TaskDto.java
@@ -2,7 +2,6 @@ package ru.mssecondteam.taskservice.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
-import ru.mssecondteam.taskservice.dto.epic.dto.EpicResponseDto;
 import ru.mssecondteam.taskservice.model.TaskStatus;
 
 import java.time.LocalDateTime;
@@ -20,6 +19,6 @@ public record TaskDto(
         Long assigneeId,
         Long authorId,
         Long eventId,
-        EpicResponseDto epic
+        Long epicId
 ) {
 }
