+++ src/main/java/ru/mssecondteam/taskservice/dto/epic/EpicResponseDto.java
@@ -1,4 +1,4 @@
-package ru.mssecondteam.taskservice.dto.epic.dto;
+package ru.mssecondteam.taskservice.dto.epic;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
