+++ src/main/java/ru/mssecondteam/taskservice/mapper/EpicMapper.java
@@ -3,9 +3,9 @@ package ru.mssecondteam.taskservice.mapper;
 import org.mapstruct.BeanMapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
-import ru.mssecondteam.taskservice.dto.epic.dto.EpicResponseDto;
-import ru.mssecondteam.taskservice.dto.epic.dto.EpicUpdateRequest;
-import ru.mssecondteam.taskservice.dto.epic.dto.NewEpicRequest;
+import ru.mssecondteam.taskservice.dto.epic.EpicResponseDto;
+import ru.mssecondteam.taskservice.dto.epic.EpicUpdateRequest;
+import ru.mssecondteam.taskservice.dto.epic.NewEpicRequest;
 import ru.mssecondteam.taskservice.model.Epic;
 
 import static org.mapstruct.NullValuePropertyMappingStrategy.IGNORE;
