+++ src/main/java/ru/mssecondteam/taskservice/service/EpicService.java
@@ -1,8 +1,7 @@
 package ru.mssecondteam.taskservice.service;
 
-import ru.mssecondteam.taskservice.dto.epic.dto.EpicUpdateRequest;
+import ru.mssecondteam.taskservice.dto.epic.EpicUpdateRequest;
 import ru.mssecondteam.taskservice.model.Epic;
-import ru.mssecondteam.taskservice.model.Task;
 
 public interface EpicService {
     Epic createEpic(Epic epic);
